requirements.txt
fastapi 
uvicorn 
redis 
langchain 
openai 
PyPDF2 
tika 
python-dotenv
email-validator
pdfminer.six
python-multipart
aiosqlite